requirements.txt
# Core dependencies
pandas==2.1.4
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Data validation
pandera==0.18.0
multimethod==1.9.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Logging
loguru==0.7.2

# Progress bars
tqdm==4.66.1